Removed v2 file upload and tests
